Move Shape and DataType into src/keras_format (2/n)
