+++ src/engine/topology.ts
@@ -17,8 +17,9 @@ import {getScopedTensorName, getUniqueTensorName, nameScope} from '../common';
 import {Constraint} from '../constraints';
 import {AttributeError, NotImplementedError, RuntimeError, ValueError} from '../errors';
 import {getInitializer, Initializer} from '../initializers';
+import {Shape} from '../keras_format/types';
 import {Regularizer} from '../regularizers';
-import {Kwargs, RegularizerFn, Shape} from '../types';
+import {Kwargs, RegularizerFn} from '../types';
 import * as generic_utils from '../utils/generic_utils';
 import * as types_utils from '../utils/types_utils';
 import * as variable_utils from '../utils/variable_utils';
