+++ src/engine/topology_test.ts
@@ -9,9 +9,11 @@
  */
 
 import {eye, memory, ones, Tensor, tensor1d, tensor2d, zeros} from '@tensorflow/tfjs-core';
+
 import * as tfl from '../index';
 import * as initializers from '../initializers';
-import {NamedTensorMap, Shape} from '../types';
+import {Shape} from '../keras_format/types';
+import {NamedTensorMap} from '../types';
 import {describeMathCPU, describeMathCPUAndGPU, expectTensorsClose} from '../utils/test_utils';
 import {LayerVariable, onesVariable, zerosVariable} from '../variables';
