+++ src/keras_format/types.ts
@@ -1,0 +1,52 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC
+ *
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ * =============================================================================
+ */
+
+/** @docalias number[] */
+export type Shape = number[];
+
+// The tfjs-core version of DataType must stay synced with this.
+export type DataType = 'float32'|'int32'|'bool'|'complex64'|'string';
+
+/**
+ * A value within the JSON-serialized form of a serializable object.
+ *
+ * The keys of any nested dicts should be in snake_case (i.e., using Python
+ * naming conventions) for compatibility with Python Keras.
+ *
+ * @see PyJsonDict
+ */
+export type PyJsonValue = boolean|number|string|null|PyJsonArray|PyJsonDict;
+
+/**
+ * A key-value dict within the JSON-serialized form of a serializable object.
+ *
+ * Serialization/deserialization uses stringified-JSON as the storage
+ * representation. Typically this should be used for materialized JSON
+ * stored on disk or sent/received over the wire.
+ *
+ * The keys of this dict and of any nested dicts should be in snake_case (i.e.,
+ * using Python naming conventions) for compatibility with Python Keras.
+ *
+ * Internally this is normally converted to a ConfigDict that has CamelCase keys
+ * (using TypeScript naming conventions) and support for Enums.
+ */
+export interface PyJsonDict {
+  [key: string]: PyJsonValue;
+}
+
+/**
+ * An array of values within the JSON-serialized form of a serializable object.
+ *
+ * The keys of any nested dicts should be in snake_case (i.e., using Python
+ * naming conventions) for compatibility with Python Keras.
+ *
+ * @see PyJsonDict
+ */
+export interface PyJsonArray extends Array<PyJsonValue> {}
