+++ src/layers/convolutional_test.ts
@@ -18,7 +18,7 @@ import {scalar, Tensor, tensor1d, tensor3d, Tensor4D, tensor4d, util} from '@ten
 import {DataFormat, PaddingMode} from '../common';
 import * as tfl from '../index';
 import {InitializerIdentifier} from '../initializers';
-import {Shape} from '../types';
+import {Shape} from '../keras_format/types';
 import {describeMathCPU, describeMathCPUAndGPU, describeMathGPU, expectTensorsClose} from '../utils/test_utils';
 
 import {conv1d, conv1dWithBias, conv2d, conv2dWithBias} from './convolutional';
