+++ src/layers/merge.ts
@@ -19,8 +19,9 @@ import {getScalar} from '../backend/state';
 import * as K from '../backend/tfjs_backend';
 import {Layer, LayerArgs, SymbolicTensor} from '../engine/topology';
 import {NotImplementedError, ValueError} from '../errors';
+import {Shape} from '../keras_format/types';
 import {l2Normalize} from '../losses';
-import {Kwargs, Shape} from '../types';
+import {Kwargs} from '../types';
 import * as generic_utils from '../utils/generic_utils';
 import * as mathUtils from '../utils/math_utils';
 import {getExactlyOneShape} from '../utils/types_utils';
